exit on handlers failure
This can sometimes happen if SendInput fails (usually with access denied).
Usually that means there was a desktop/session change that went unnoticed
by the helper service (reasons still unclear). Exiting in that state means
wga will be restarted by the service on the proper desktop/session.

+++ include/handlers.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <windows.h>
 
-ULONG handle_xconf(void);
-ULONG handle_server_data(void);
+DWORD handle_xconf(void);
+DWORD handle_server_data(void);
